Make sure offending files are logged
